agrof-main/src/api/requirements_lightweight.txt
# Lightweight Production Requirements for AGROF
# Optimized for 512MB RAM deployment

# Core web framework
flask>=2.0.0
flask-cors>=3.0.10
gunicorn>=20.1.0

# HTTP requests
requests>=2.28.0

# Environment variables
python-dotenv>=0.19.0

# Image processing (lightweight)
Pillow>=9.0.0
numpy>=1.21.0

# Google Cloud Vision API (lightweight)
google-cloud-vision>=3.4.4

# Basic machine learning (CPU only, lightweight)
tensorflow-cpu>=2.13.0

# Computer vision (lightweight)
opencv-python-headless>=4.8.0.76

# Data processing (lightweight)
pandas>=2.0.3

# Monitoring and logging
loguru>=0.7.0
